Move onoff server into codegen subfolder.
