+++ src/app/clusters/on-off-server/codegen/scenes-integration.cpp
